:adhesive_bandage: Update Warzone model to use BigInteger

+++ models.py
@@ -1,6 +1,6 @@
 from os import getenv
 
-from sqlalchemy import Boolean, Column, Integer, String, create_engine
+from sqlalchemy import Boolean, Column, BigInteger, String, create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.sql.sqltypes import Boolean
@@ -16,8 +16,8 @@ session = Session()
 class Warzone(Base):
     __tablename__ = "warzone"
     battletag = Column(String, primary_key=True)
-    channel_id = Column(Integer, nullable=True)
-    member_id = Column(Integer, nullable=False)
+    channel_id = Column(BigInteger, nullable=True)
+    member_id = Column(BigInteger, nullable=False)
     track = Column(Boolean, nullable=False, default=False)
     last_match = Column(String, nullable=True)
